+++ src/configDescription.ts
@@ -230,10 +230,19 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     ],
   ],
   [
-    "fastBossRush",
+    "removeArmor",
     [
       ModConfigMenuOptionType.BOOLEAN,
       "0302",
+      "Remove armor",
+      "Remove damage scaling from all of the bosses that have it.",
+    ],
+  ],
+  [
+    "fastBossRush",
+    [
+      ModConfigMenuOptionType.BOOLEAN,
+      "0303",
       "Fast Boss Rush",
       "The mandatory waiting between bosses is removed.",
     ],
@@ -242,7 +251,7 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     "killExtraEnemies", // 45, 78
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0303",
+      "0304",
       "Kill extra enemies",
       "Extra enemies will properly die after defeating Mom, Mom's Heart, or It Lives!",
     ],
@@ -251,7 +260,7 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     "fastPin", // 62
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0304",
+      "0305",
       "Fast Pin",
       "Make Pin, Frail, Scolex, and Wormwood spend less time underground.",
     ],
@@ -260,7 +269,7 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     "preventDeathSlow", // 66
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0305",
+      "0306",
       "Stop Death's slow attack",
       "Stop Death from performing the attack that reduces your speed by a factor of 2.",
     ],
@@ -269,7 +278,7 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     "fastKrampus", // 81
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0306",
+      "0307",
       "Fast Krampus",
       "Make Krampus immediately drop his item once he is killed.",
     ],
@@ -278,7 +287,7 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     "fastSatan", // 84
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0307",
+      "0308",
       "Fast Satan",
       "All of the waiting during the Satan Fight is removed.",
     ],
@@ -287,7 +296,7 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     "fastHaunt", // 260
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0308",
+      "0309",
       "Fast The Haunt",
       "Some animations in The Haunt fight are sped up.",
     ],
@@ -296,18 +305,9 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
     "fastAngels", // 271, 272
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0309",
+      "0310",
       "Fast angels",
       "Make Uriel and Gabriel immediately drop their key piece once they are killed.",
-    ],
-  ],
-  [
-    "consistentAngels", // 271, 272
-    [
-      ModConfigMenuOptionType.BOOLEAN,
-      "0310",
-      "Consistent angels",
-      "Prevent two of the same angel boss from spawning in the same room.",
     ],
   ],
 ];
@@ -315,10 +315,19 @@ export const BOSS_CHANGES_1: ConfigDescriptions = [
 // 0311-0320
 export const BOSS_CHANGES_2: ConfigDescriptions = [
   [
-    "removeLambBody", // 273
+    "consistentAngels", // 271, 272
     [
       ModConfigMenuOptionType.BOOLEAN,
       "0311",
+      "Consistent angels",
+      "Prevent two of the same angel boss from spawning in the same room.",
+    ],
+  ],
+  [
+    "removeLambBody", // 273
+    [
+      ModConfigMenuOptionType.BOOLEAN,
+      "0312",
       "Remove The Lamb body",
       "Remove The Lamb body entirely once it dies.",
     ],
@@ -327,18 +336,9 @@ export const BOSS_CHANGES_2: ConfigDescriptions = [
     "preventVictoryLapPopup", // 273
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0312",
+      "0313",
       "Stop the Victory Lap popup",
       "Prevent the Victory Lap popup from appearing once you defeat The Lamb.",
-    ],
-  ],
-  [
-    "removeMegaSatanArmor", // 275
-    [
-      ModConfigMenuOptionType.TEXT,
-      "0313",
-      "Remove Mega Satan's armor",
-      "Remove Mega Satan's armor during his second phase, which is intended to make it scale with powerful builds.",
     ],
   ],
   [
@@ -351,19 +351,10 @@ export const BOSS_CHANGES_2: ConfigDescriptions = [
     ],
   ],
   [
-    "removeHushArmor", // 407
-    [
-      ModConfigMenuOptionType.BOOLEAN,
-      "0315",
-      "Remove Hush Armor",
-      "Remove Hush's armor, which is intended to make it scale with powerful builds.",
-    ],
-  ],
-  [
     "fastBigHorn", // 411
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0316",
+      "0315",
       "Fast Big Horn",
       "Make Big Horn spend less time underground.",
     ],
@@ -372,7 +363,7 @@ export const BOSS_CHANGES_2: ConfigDescriptions = [
     "fastColostomia", // 917
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0317",
+      "0316",
       "Fast Colostomia",
       "Make Colostomia appear instantly.",
     ],
@@ -381,9 +372,9 @@ export const BOSS_CHANGES_2: ConfigDescriptions = [
     "fastDogma", // 950
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "0318",
+      "0317",
       "Fast Dogma",
-      "Automatically skip Dogma's death cutscene.",
+      "Speed up Dogma's death animation (and skip the death cutscene).",
     ],
   ],
   [
@@ -404,10 +395,6 @@ export const BOSS_CHANGES_2: ConfigDescriptions = [
       "Defeating Mega Satan no longer has a chance to immediately end the run.",
     ],
   ],
-];
-
-// 0311-0320
-export const BOSS_CHANGES_3: ConfigDescriptions = [
   [
     "", // 407
     [
@@ -417,6 +404,10 @@ export const BOSS_CHANGES_3: ConfigDescriptions = [
       'Make Hush no longer play an "Appear" animation.',
     ],
   ],
+];
+
+// 0311-0320
+export const BOSS_CHANGES_3: ConfigDescriptions = [
   [
     "", // 951
     [