+++ src/features/optional/bosses/removeArmor.ts
@@ -4,6 +4,30 @@ import { config } from "../../../modConfigMenu";
 let dealingManualDamage = false;
 
 // ModCallback.ENTITY_TAKE_DMG (11)
+// EntityType.MEGA_SATAN_2 (275)
+export function entityTakeDmgMegaSatan2(
+  entity: Entity,
+  amount: float,
+  damageFlags: BitFlags<DamageFlag>,
+  source: EntityRef,
+  countdownFrames: int,
+): boolean | undefined {
+  if (!config.removeArmor) {
+    return;
+  }
+
+  if (dealingManualDamage) {
+    return;
+  }
+
+  dealingManualDamage = true;
+  entity.TakeDamage(amount, damageFlags, source, countdownFrames);
+  dealingManualDamage = false;
+
+  return false;
+}
+
+// ModCallback.ENTITY_TAKE_DMG (11)
 // EntityType.HUSH (407)
 export function entityTakeDmgHush(
   entity: Entity,
@@ -12,7 +36,7 @@ export function entityTakeDmgHush(
   source: EntityRef,
   countdownFrames: int,
 ): boolean | undefined {
-  if (!config.removeHushArmor) {
+  if (!config.removeArmor) {
     return;
   }